pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pretrain-experiments"
version = "0.1.0"
description = "A research framework for controlled pretraining experiments on language models"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Sebastian Bordt"}
]
keywords = ["pretraining", "language-models", "olmo", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch",
    "transformers",
    "wandb",
    "numpy",
    "pyyaml",
    "tqdm",
    "huggingface_hub",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "ruff",
]

[project.scripts]
pretrain-experiments = "pretrain_experiments.cli:main"

[project.urls]
Repository = "https://github.com/sbordt/pretrain-experiments"

[tool.setuptools.packages.find]
where = ["."]
include = ["pretrain_experiments*"]

[tool.setuptools.package-data]
pretrain_experiments = ["*.yaml", "*.json"]
